refactor: 파일 위치 수정 및 주석 삭제
