+++ src/main/java/spharos/payment/application/PaymentServiceImpl.java
@@ -11,9 +11,7 @@ import org.axonframework.commandhandling.gateway.CommandGateway;
 import org.springframework.stereotype.Service;
 import spharos.payment.axon.command.SavePaymentCommand;
 import spharos.payment.batch.BatchScheduler;
-import spharos.payment.domain.Payment;
-import spharos.payment.domain.PaymentStatus;
-import spharos.payment.domain.PaymentType;
+import spharos.payment.domain.enumPackage.PaymentStatus;
 import spharos.payment.dto.PaymentRequest;
 import spharos.payment.dto.PaymentResultResponse;
 import spharos.payment.infrastructure.PaymentRepository;
@@ -26,13 +24,10 @@ public class PaymentServiceImpl  {
 
     private final PaymentRepository paymentRepository;
     private final CommandGateway commandGateway;
-
     private final BatchScheduler batchScheduler;
+
     //결제 저장
     public void savePayment(PaymentRequest paymentRequest) {
-   /*     Payment payment = Payment.createPayment(paymentRequest.getClientEmail(),  paymentRequest.getPayType(),
-                paymentRequest.getTotalAmount(), paymentRequest.getApprovedAt(),paymentRequest.getPayStatus());
-        paymentRepository.save(payment);*/
         SavePaymentCommand savePaymentCommand = new SavePaymentCommand(UUID.randomUUID().toString(),
                 paymentRequest.getClientEmail(),
                 paymentRequest.getPayType(), paymentRequest.getTotalAmount(), paymentRequest.getApprovedAt(),
@@ -43,15 +38,10 @@ public class PaymentServiceImpl  {
     public List<PaymentResultResponse> getPaymentsList(){
         LocalDate requestDate = LocalDate.now();
 
-        LocalDateTime parse = LocalDateTime.parse(requestDate + "T00:00:00");
-        LocalDateTime parse1 = LocalDateTime.parse(requestDate + "T23:59:58");
+        LocalDateTime startTime = LocalDateTime.parse(requestDate + "T00:00:00");
+        LocalDateTime endTime = LocalDateTime.parse(requestDate + "T23:59:58");
 
-        log.info("firstDayOfMonth : {}", parse);
-        log.info("lastDayOfMonth : {}", parse1);
-        var doneStatus = PaymentStatus.DONE;
-        var cancelStatus = PaymentStatus.CANCEL;
-
-        List<PaymentResultResponse> payments = paymentRepository.findByApprovedAtAndPaymentStatus(parse, parse1);
+        List<PaymentResultResponse> payments = paymentRepository.findByApprovedAtAndPaymentStatus(startTime, endTime);
         return payments;
 
     }
@@ -59,8 +49,7 @@ public class PaymentServiceImpl  {
     //배치 하고 db에서 이름만 뽑아서 이름으로 밸류 찾고 그걸 카프카로 보내보기
     public void batchTest(){
         batchScheduler.runJob();
-        //List<Payment> all = paymentRepository.findAll();
-       // return all;
+
     }
