+++ src/main/java/spharos/payment/axon/agegregate/PaymentAggregate.java
@@ -5,14 +5,11 @@ import static org.axonframework.modelling.command.AggregateLifecycle.apply;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.axonframework.commandhandling.CommandHandler;
-import org.axonframework.eventhandling.EventHandler;
 import org.axonframework.eventsourcing.EventSourcingHandler;
 import org.axonframework.modelling.command.AggregateIdentifier;
 import org.axonframework.spring.stereotype.Aggregate;
 import spharos.payment.axon.command.SavePaymentCommand;
 import spharos.payment.axon.event.PaymentSaveEvent;
-import spharos.payment.domain.PaymentStatus;
-import spharos.payment.domain.PaymentType;
 
 
 @Aggregate
@@ -24,6 +21,7 @@ public class PaymentAggregate {
     private String id;
 
 
+    // 결제 저장
     @CommandHandler
     public PaymentAggregate(SavePaymentCommand command)  {
         //throw new RuntimeException("Custom runtime exception message");
