+++ src/main/java/spharos/payment/axon/event/PaymentSaveEvent.java
@@ -4,8 +4,6 @@ import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import spharos.payment.domain.PaymentStatus;
-import spharos.payment.domain.PaymentType;
 
 
 @AllArgsConstructor
