+++ src/main/java/spharos/payment/dto/BatchPayment.java
@@ -1,15 +1,8 @@
 package spharos.payment.dto;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Convert;
-import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import spharos.payment.domain.PaymentStatus;
-import spharos.payment.domain.PaymentStatusConverter;
-import spharos.payment.domain.PaymentType;
-import spharos.payment.domain.PaymentTypeConverter;
 
 @Data
 @AllArgsConstructor
